feat(Conflicts): Move to `Data.Conflicts`
